refactor: Move item package to internal pkg
