+++ cmd/convert.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/glass-cms/glasscms/item"
+	"github.com/glass-cms/glasscms/internal/item"
 	"github.com/glass-cms/glasscms/parser"
 	"github.com/glass-cms/glasscms/sourcer"
 	"github.com/lmittmann/tint"
