+++ parser/parser.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/glass-cms/glasscms/item"
+	"github.com/glass-cms/glasscms/internal/item"
 	"github.com/glass-cms/glasscms/sourcer"
 	"github.com/mozillazg/go-slugify"
 	"gopkg.in/yaml.v3"
